Move module tests to pytest
